results/rebuttal_new_baselines/run-20231112-235354_LLMatic-ZeroShot_donnemartin-system-design-oop-finance2_0_20-runs_log/finance2/17/LLMatic/expense_income.py
class Expense:
	def __init__(self):
		self.expenses = {}

	def import_expenses(self, expenses):
		for expense in expenses:
			self.expenses[expense['name']] = expense['amount']

	def categorize_expenses(self, categories):
		self.expenses = {}
		for category in categories:
			self.expenses[category['name']] = category['amount']

	def visualize_expenses(self):
		return self.expenses


class Income:
	def __init__(self):
		self.incomes = {}

	def import_incomes(self, incomes):
		for income in incomes:
			self.incomes[income['name']] = income['amount']

	def categorize_incomes(self, categories):
		self.incomes = {}
		for category in categories:
			self.incomes[category['name']] = category['amount']

	def visualize_incomes(self):
		return self.incomes
